Refactoring utils test and utils hepler code structure
